requirements.txt
aiogram
networkx
numpy
pyyaml
tokenize_uk
